// src/assets/types.ts
export enum IconsType {
  down = 'down',
  filter = 'filter',
  backlog = 'backlog',
  todo = 'todo',
  inProgress = 'inProgress',
  done = 'done',
  cancelled = 'cancelled',
  high = 'high',
  low = 'low',
  medium = 'medium',
  urgent = 'urgent',
  plus = 'plus',
  more = 'more',
  noPriority = 'noPriority',
}